requirements.txt
alembic==1.15.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crcmod==1.7
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
dotenv==0.9.9
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
importlib_metadata==8.6.1
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
load-dotenv==0.1.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
oss2==2.19.1
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
yarg==0.1.9
zipp==3.21.0
